Create What's that data type?
Do you remember that when you added 5 + "six", you got an error due to a mismatch in data types? You can avoid such embarrassing situations by checking the data type of a variable beforehand. You can do this with the class() function, as the code in the editor shows.


Complete the code in the editor and also print out the classes of my_character and my_logical.