Create Battle of the sexes
You might wonder what happens when you try to compare elements of a factor. In factor_survey_vector you have a factor with two levels: "Male" and "Female". But how does R value these relative to each other?

Instructions

Read the code in the editor and submit the answer to test if male is greater than (>) female.